+++ crates/sncast/src/starknet_commands/multicall/invoke.rs
@@ -33,7 +33,6 @@ impl MulticallInvoke {
                 },
             },
         };
-
         Ok(invoke)
     }