+++ crates/sncast/src/starknet_commands/multicall/mod.rs
@@ -1,15 +1,18 @@
-use anyhow::Result;
+use anyhow::{Context, Result};
 use clap::{Args, Subcommand};
 use serde::Serialize;
 use serde_json::{Value, json};
 
 pub mod contract_registry;
 pub mod deploy;
+pub mod execute;
 pub mod invoke;
 pub mod new;
 pub mod run;
 
 use crate::starknet_commands::multicall::contract_registry::ContractRegistry;
+use crate::starknet_commands::multicall::execute::Execute;
+use crate::starknet_commands::utils::felt_or_id::FeltOrId;
 use crate::{Arguments, process_command_result, starknet_commands};
 use foundry_ui::Message;
 use new::New;
@@ -34,6 +37,7 @@ pub struct Multicall {
 pub enum Commands {
     Run(Box<Run>),
     New(New),
+    Execute(Box<Execute>),
 }
 
 pub async fn multicall(
@@ -96,34 +100,60 @@ pub async fn multicall(
             process_command_result("multicall run", result, ui, block_explorer_link);
             Ok(())
         }
+        starknet_commands::multicall::Commands::Execute(execute) => {
+            let provider = execute.rpc.get_provider(&config, ui).await?;
+            let account = get_account(&config, &provider, &execute.rpc, ui).await?;
+
+            let result = with_account!(&account, |account| {
+                starknet_commands::multicall::execute::execute(
+                    *execute.clone(),
+                    account,
+                    &provider,
+                    wait_config,
+                    ui,
+                )
+                .await
+            });
+            let block_explorer_link =
+                block_explorer_link_if_allowed(&result, provider.chain_id().await?, &config).await;
+            process_command_result("multicall execute", result, ui, block_explorer_link);
+            Ok(())
+        }
     }
 }
 
 /// Replaces arguments that reference user-defined ids with their corresponding values from the contract registry.
 pub fn replaced_arguments(
     arguments: &Arguments,
-    contract_registry: &ContractRegistry,
+    contracts: &ContractRegistry,
 ) -> Result<Arguments> {
     Ok(match (&arguments.calldata, &arguments.arguments) {
         (Some(calldata), None) => {
             let replaced_calldata = calldata
                 .iter()
-                .map(|input| {
-                    if let Some(address) = contract_registry.get_address_by_id(input) {
-                        Ok(address.to_string())
+                .map(|raw_input| {
+                    let input = FeltOrId::new(raw_input.clone());
+                    if let Some(id) = input.as_id() {
+                        contracts
+                            .get_address_by_id(id)
+                            .with_context(|| {
+                                format!("Failed to find contract address for id: {id}")
+                            })
+                            .map(|f| f.to_string())
                     } else {
-                        Ok(input.clone())
+                        Ok(raw_input.clone())
                     }
                 })
                 .collect::<Result<Vec<String>>>()?;
+
             Arguments {
                 calldata: Some(replaced_calldata),
                 arguments: None,
             }
         }
         (None, _) => arguments.clone(),
-        (Some(_), Some(_)) => anyhow::bail!(
-            "Invalid arguments: both `calldata` and `arguments` are set. Please provide only one."
-        ),
+        (Some(_), Some(_)) => {
+            unreachable!("Only one of `calldata` or `arguments` can be provided, but both are set.")
+        }
     })
 }